test(tools): make the subagent_type backfill tests actually run
The regression tests added for the `subagent_type` backfill landed in
`tests/group_a.rs`, which is commented out of the module tree
(`tests.rs`, `// mod group_a;`) because it still references the
renamed-and-now-private `parse_minimax_xml_call`. As a result the two
new tests never compiled and never ran — `cargo test` was green but
exercised nothing, and `group_a.rs` was the only place `backfill_*` was
covered at all. The +90 lines also pushed `group_a.rs` from 460 to 550
LoC, a fresh violation of the ≤500-LoC-per-file cap.

Relocate the two tests verbatim into a new LIVE module
`tests/group_e.rs` (registered via `mod group_e;`) and revert
`group_a.rs` to its prior state. The tests now compile and pass
(2 passed), give real coverage to `backfill_required_params` +
`validate_single_tool_call` for delegation tools, and no file crosses
the LoC cap. The pre-existing disabled-`group_a` TODO is left untouched
(out of scope).

The fix under test (validation.rs `infer_default_subagent_type`) is
unchanged.
