+++ crates/spark-server/src/tool_parser/tests.rs
@@ -8,3 +8,4 @@
 mod group_b;
 mod group_c;
 mod group_d;
+mod group_e;
